🚚 rename(#7): CourseCommandController 패키지 이동
